Change dashboard templates to be 1 item (#3195)
* Put templates under 1 item

* Upgrade dependencies

* Fix weird apollo caching bug

* Fix type error

* Remove need for teamId in unmakeTemplates

* Fix cache update function
