+++ packages/app/package.json
@@ -70,7 +70,7 @@
     ]
   },
   "dependencies": {
-    "@apollo/react-hooks": "^3.0.1",
+    "@apollo/react-hooks": "^3.1.3",
     "@babel/plugin-transform-destructuring": "^7.5.0",
     "@babel/preset-env": "^7.5.5",
     "@codesandbox/executors": "^0.1.0",
@@ -83,7 +83,7 @@
     "@types/rc-slider": "^8.6.5",
     "@vue/babel-preset-app": "^3.2.0",
     "airtable": "^0.5.8",
-    "apollo-boost": "^0.4.3",
+    "apollo-boost": "^0.4.4",
     "apollo-link-batch-http": "^1.2.12",
     "apollo-link-context": "^1.0.18",
     "axios": "^0.19.0",
