+++ packages/app/src/app/graphql/schema.graphql
@@ -1,5 +1,5 @@
-# source: https://codesandbox.io/api/graphql
-# timestamp: Wed Dec 11 2019 16:37:29 GMT+0100 (Central European Standard Time)
+# source: https://codesandbox.stream/api/graphql
+# timestamp: Thu Dec 12 2019 17:24:06 GMT+0100 (Central European Standard Time)
 
 schema {
   query: RootQuery
