+++ packages/app/src/app/pages/Dashboard/Content/index.js
@@ -3,7 +3,7 @@ import { Route, Switch, Redirect, withRouter } from 'react-router-dom';
 
 import { RecentSandboxes } from './routes/RecentSandboxes';
 import PathedSandboxes from './routes/PathedSandboxes';
-import { Templates, FollowedTemplates } from './routes/Templates';
+import { Templates } from './routes/Templates';
 import DeletedSandboxes from './routes/DeletedSandboxes';
 import SearchSandboxes from './routes/SearchSandboxes';
 import CreateTeam from './routes/CreateTeam';
@@ -14,7 +14,6 @@ const Content = () => (
     <Route path="/dashboard/recent" component={RecentSandboxes} />
     <Route path="/dashboard/trash" component={DeletedSandboxes} />
     <Route path="/dashboard/templates" exact component={Templates} />
-    <Route path="/dashboard/templates/followed" component={FollowedTemplates} />
     <Route path="/dashboard/sandboxes/:path*" component={PathedSandboxes} />
     <Route path="/dashboard/search" component={SearchSandboxes} />
     <Route path="/dashboard/teams/new" component={CreateTeam} />
@@ -27,10 +26,6 @@ const Content = () => (
       path="/dashboard/teams/:teamId/templates"
       component={Templates}
       exact
-    />
-    <Route
-      path="/dashboard/teams/:teamId/templates/followed"
-      component={FollowedTemplates}
     />
     <Redirect to="/dashboard/recent" />
   </Switch>
