+++ packages/app/src/app/pages/Dashboard/Content/routes/Templates/OwnedTemplates/OwnedTemplates.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Helmet } from 'react-helmet';
 import { sortBy } from 'lodash-es';
 import { useQuery } from '@apollo/react-hooks';
 
@@ -20,24 +19,23 @@ import {
   ListTemplatesQuery,
   ListTemplatesQueryVariables,
 } from 'app/graphql/types';
-import { RouteComponentProps } from 'react-router';
-import { Container, Grid, EmptyTitle } from '../elements';
+import { Grid, EmptyTitle } from '../elements';
 import { Navigation } from '../Navigation';
 
-type TemplatesProps = RouteComponentProps<{ teamId: string }> & {};
+type OwnedTemplatesProps = { teamId?: string };
 
-export const Templates = (props: TemplatesProps) => {
+export const OwnedTemplates = (props: OwnedTemplatesProps) => {
   const {
     actions: {
       dashboard: { deleteTemplate },
     },
   } = useOvermind();
-  const { teamId } = props.match.params;
+  const { teamId } = props;
   const { loading, error, data, refetch } = useQuery<
     ListTemplatesQuery,
     ListTemplatesQueryVariables
   >(LIST_OWNED_TEMPLATES, {
-    variables: { teamId, showAll: false },
+    variables: { showAll: false },
   });
 
   if (error) {
@@ -56,20 +54,21 @@ export const Templates = (props: TemplatesProps) => {
           color: 'rgba(255, 255, 255, 0.5)',
         }}
       >
-        Fetching Sandboxes...
+        Fetching Templates...
       </DelayedAnimation>
     );
   }
 
-  const sortedTemplates = sortBy(data.me.templates, template =>
+  const templatesToUse = teamId
+    ? data.me.teams.find(t => t.id === teamId)?.templates
+    : data.me.templates;
+
+  const sortedTemplates = sortBy(templatesToUse, template =>
     getSandboxName(template.sandbox).toLowerCase()
   );
 
   return (
-    <Container>
-      <Helmet>
-        <title>{teamId ? 'Team' : 'My'} Templates - CodeSandbox</title>
-      </Helmet>
+    <>
       <Navigation teamId={teamId} number={sortedTemplates.length} />
       {!sortedTemplates.length && (
         <div>
@@ -96,7 +95,7 @@ export const Templates = (props: TemplatesProps) => {
                   title: 'Convert to Sandbox',
                   action: () => {
                     track('Template - Removed', { source: 'Context Menu' });
-                    unmakeTemplates([template.sandbox.id], teamId);
+                    unmakeTemplates([template.sandbox.id]);
                     return true;
                   },
                 },
@@ -123,6 +122,6 @@ export const Templates = (props: TemplatesProps) => {
           </ContextMenu>
         ))}
       </Grid>
-    </Container>
+    </>
   );
 };
